fix: component validation
